Switched Document#isNewNote to use Long.valueOf instead of Integer.valueOf. This allows saving of "high bit" design note IDs like "FFFF0002" for the About document

+++ org.openntf.domino/src/org/openntf/domino/impl/Document.java
@@ -1806,7 +1806,7 @@ public class Document extends Base<org.openntf.domino.Document, lotus.domino.Doc
 	 */
 	@Override
 	public boolean isNewNote() {
-		return Integer.valueOf(noteid_, 16) == 0;
+		return Long.valueOf(noteid_, 16) == 0;
 	}
 
 	/*
